restructure directories for setups for specific pages
