+++ src/components/setups/index.ts
@@ -1,0 +1,7 @@
+export * from './general'
+export * from './navbar'
+export * from './home'
+export * from './about'
+export * from './projects'
+export * from './experience'
+export * from './contact'
